requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic-settings==2.2.1
asyncpg==0.29.0
supabase==2.4.0
pyjwt==2.8.0
httpx==0.27.0
tiktoken==0.6.0
slowapi==0.1.9
python-dotenv==1.0.1
# For dev/testing
pytest==7.4.3
pytest-asyncio==0.21.1
